libc: implement more standard headers, improve architecture splitting

+++ store/libc/lib/string/string_ext.c
@@ -2,7 +2,7 @@
 
 size_t strnlen(const char *s, size_t n) {
 	const char *p = memchr(s, 0, n);
-	return p ? p - s : n;
+	return p ? (size_t) (p - s) : n;
 }
 
 char* stpcpy(char* restrict d, const char* restrict s) {
